requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Security
bcrypt==4.1.2
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Reports (optional, for PDF/Excel export)
reportlab==4.0.7
openpyxl==3.1.2
